See Desc.
Okay, I tried once to implement dynamic (un)loading of renderer apis, but I couldn't get premake to not include said renderer api dll projects in the build if they weren't supported. As in, I can't use the include command in a filter context and have it only be included in that context; the example being to not include the renderer apis on platforms they're not supported on.
I'll try again tomorrow, which is why I left in some framework for it, like WindowsRendererAPILoader.cpp.